Moving Secrets functionality to OpensSearchSecrets
