+++ tests/unit/lib/test_opensearch_secrets.py
@@ -1,9 +1,9 @@
 # Copyright 2023 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-import unittest
 from unittest.mock import MagicMock, patch
 
+import pytest
 from charms.opensearch.v0.constants_charm import (
     ClientRelationName,
     PeerRelationName,
@@ -11,13 +11,22 @@ from charms.opensearch.v0.constants_charm import (
 )
 from charms.opensearch.v0.constants_tls import CertType
 from charms.opensearch.v0.opensearch_base_charm import SERVICE_MANAGER
+from ops import JujuVersion
 from ops.testing import Harness
+from overrides import override
+from parameterized import parameterized
 
 from charm import OpenSearchOperatorCharm
 
+from .test_opensearch_internal_data import TestOpenserchInternalData
 
-class TestOpenSearchSecrets(unittest.TestCase):
-    """Testing OpenSearchSecrets component."""
+
+@pytest.mark.usefixtures("only_with_juju_secrets")
+class TestOpenSearchSecrets(TestOpenserchInternalData):
+    """Ensuring that secrets interfaces and expected behavior are preserved.
+
+    Additionally the class also highlights the difference introdced in SecretsDataStore
+    """
 
     def setUp(self):
         self.harness = Harness(OpenSearchOperatorCharm)
@@ -28,6 +37,7 @@ class TestOpenSearchSecrets(unittest.TestCase):
         self.app = self.charm.app
         self.unit = self.charm.unit
         self.secrets = self.charm.secrets
+        self.secret_store = self.charm.secrets
 
         self.peers_rel_id = self.harness.add_relation(PeerRelationName, self.charm.app.name)
         self.service_rel_id = self.harness.add_relation(SERVICE_MANAGER, self.charm.app.name)
@@ -84,3 +94,79 @@ class TestOpenSearchSecrets(unittest.TestCase):
         self.secrets.put_object(scope, "obj", {"key1": "val1"})
         self.assertTrue(self.secrets.has(scope, "obj"))
         self.assertTrue(self.secrets.get_object(scope, "obj"), {"key1": "val1"})
+
+    def test_implements_secrets(self):
+        """Property determining whether secerts are available."""
+        self.assertEqual(
+            self.secret_store.implements_secrets, JujuVersion.from_environ().has_secrets
+        )
+
+    @override
+    @parameterized.expand([(Scope.APP), (Scope.UNIT)])
+    def test_put_get_set_object_implementation_specific_behavior(self, scope):
+        """Test putting and getting objects in/from the secret store."""
+        self.secret_store.put_object(scope, "key-obj", {"name1": "val1"}, merge=True)
+        self.secret_store.put_object(
+            scope, "key-obj", {"name1": None, "name2": "val2"}, merge=True
+        )
+        self.assertDictEqual(self.secret_store.get_object(scope, "key-obj"), {"name2": "val2"})
+
+    @override
+    @parameterized.expand([(Scope.APP), (Scope.UNIT)])
+    def test_nullify_obj(self, scope):
+        """Test iteratively filling up an object with `None` values."""
+        self.secret_store.put_object(scope, "key-obj", {"key1": "val1", "key2": "val2"})
+        self.secret_store.put_object(scope, "key-obj", {"key1": None, "key2": "val2"}, merge=True)
+        self.secret_store.put_object(scope, "key-obj", {"key2": None}, merge=True)
+        self.assertFalse(self.secret_store.has(scope, "key-obj"))
+
+    def test_label_app(self):
+        scope = Scope.APP
+        label = self.secret_store.label(scope, "key1")
+        self.assertEqual(label, f"opensearch:{scope}:key1")
+        self.assertEqual(
+            self.secret_store.breakdown_label(label),
+            {"application_name": "opensearch", "scope": scope, "unit_id": None, "key": "key1"},
+        )
+
+    def test_label_unit(self):
+        scope = Scope.UNIT
+        label = self.secret_store.label(scope, "key1")
+        self.assertEqual(self.secret_store.label(scope, "key1"), f"opensearch:{scope}:0:key1")
+        self.assertEqual(
+            self.secret_store.breakdown_label(label),
+            {"application_name": "opensearch", "scope": scope, "unit_id": 0, "key": "key1"},
+        )
+
+    @parameterized.expand([(Scope.APP), (Scope.UNIT)])
+    def test_save_secret_id(self, scope):
+        """Test putting and getting objects in/from the secret store."""
+        self.secret_store.put(scope, "key", "val1")
+        secret_id = self.secret_store._get_relation_data(scope)[
+            self.secret_store.label(scope, "key")
+        ]
+        secret_content = self.charm.model.get_secret(id=secret_id).get_content()
+        self.assertEqual(secret_content["key"], "val1")
+
+        self.secret_store.put_object(scope, "key-obj", {"name1": "val1"}, merge=True)
+        secret_id2 = self.secret_store._get_relation_data(scope)[
+            self.secret_store.label(scope, "key-obj")
+        ]
+        secret_content = self.charm.model.get_secret(id=secret_id2).get_content()
+        self.assertEqual(secret_content["name1"], "val1")
+
+    def test_bad_label(self):
+        with self.assertRaises(IndexError):
+            self.secret_store.breakdown_label("bla")
+
+        with self.assertRaises(IndexError):
+            self.secret_store.breakdown_label("bla-bla-bla")
+
+        with self.assertRaises(UnboundLocalError):
+            self.secret_store.breakdown_label("bla:bla")
+
+        with self.assertRaises(UnboundLocalError):
+            self.secret_store.breakdown_label("bla:bla:bla")
+
+        with self.assertRaises(ValueError):
+            self.secret_store.breakdown_label("bla:bla:bla:bla")
